Update JS path in umbraco-package.json entrypoints
Simplified the path to the JavaScript file in the `entrypoints`
section of `umbraco-package.json` for the entrypoint
`Tollervey.Umbraco.LightningPayments.Entrypoint`. The previous
absolute path `"/App_Plugins/Tollervey.Umbraco.LightningPayments/
lightning-ui.js"` was replaced with the relative path
`"lightning-ui.js"`. This reflects a change in file location or
reference method.