Remove `<StaticWebAssetBasePath>` from project file
The `<StaticWebAssetBasePath>` property was removed from the
`Tollervey.Umbraco.LightningPayments.UI.csproj` file. This property
was previously used to specify the base path for static web assets
in the project. Its removal suggests that the default behavior for
serving static assets is now sufficient, or that static assets are
no longer being served in this manner.