Prevent duplicate manifest registration, update conditions
Added a check in `entrypoint.ts` to prevent duplicate manifest
registrations in the `extensionRegistry`, logging a warning
and skipping registration if duplicates are found.

Updated `conditions` in `manifests.ts` to use `match` instead
of `value` for consistency. Modified `defaultData` in
`propertyEditorManifests` to use `!1` as shorthand for `false`.

Improved comments for clarity, including notes on path
adjustments and the rationale for `conditions` updates.